Ms new design failed

+++ src/mlmc/sampling_pool_pbs.py
@@ -7,7 +7,43 @@ import json
 import glob
 from mlmc.level_simulation import LevelSimulation
 from mlmc.sampling_pool import SamplingPool
-from mlmc.pbs_process import PbsProcess
+from mlmc.pbs_job import PbsJob
+
+
+"""
+SamplingPoolPBS description
+    - this class inherits from SampleStorage, both abstract methods and other crucial ones are described
+
+    schedule_sample(sample_id, level_sim)
+        - serialize level_sim (mlmc/level_simulation.py), pickle is used
+        - compute random seed from sample_id
+        - add (level_sim.level_id, sample_id, seed) to job's scheduled samples 
+        - add job weight, increment number of samples in job and execute if job_weight is exceeded
+    
+    execute()
+        - it is call when job weight (Maximum sum of task sizes summation in single one job) is exceeded
+        - methods from mlmc/pbs_job.py are called
+            - PbsJob class is created and serialized (PbsJob static method does both)
+            - scheduled samples are saved through PbsJob class static method
+        - pbs script is written out and ready to run
+        
+    get_finished()
+        - run execute()
+        - get finished_pbs_jobs and unfinished_pbs_jobs from qstat output
+        - call get_result_files(), it returns successful samples, failed samples and times all of that is return to Sampler
+        
+    _get_result_files()
+        - set n_running - number of running samples, it is given from unfinished_pbs_jobs
+        - successful samples, failed samples and run times are retrieved from PbsJob class with given job_id
+        - if there are _unfinished_sample_ids ('renew' command was use) these samples are appended to previous ones
+        
+    
+    This class cooperate with PbsJob (mlmc/pbs_job), which is used as "mediator" between master process and
+    worker (job) process. Data which are necessary for worker process are passed to PbsJob from SampleStoragePbs. 
+    Master process serializes PbsJob instance.
+    Then PbsJob is deserialized in worker process.
+    
+"""
 
 
 class SamplingPoolPBS(SamplingPool):
@@ -16,12 +52,13 @@ class SamplingPoolPBS(SamplingPool):
     JOBS_DIR = "jobs"
     LEVEL_SIM_CONFIG = "level_{}_simulation_config"  # Serialized level simulation
     JOB = "{}_job.sh"  # Pbs process file
+    JOBS_COUNT = "jobs_count.txt" # Contains current number of jobs which is also job unique identifier
 
-    def __init__(self, work_dir, job_weight=200000, job_count=0):
+    def __init__(self, work_dir, job_weight=200000, clean=False):
         """
         :param work_dir: Path to working directory
         :param job_weight: Maximum sum of task sizes summation in single one job, if this value is exceeded then the job is executed
-        :param job_count: Number of created jobs, it holds the current jobID, useful for regenerating of failed samples (jobs)
+        :param clean: bool, if True delete output dir
         """
         self._work_dir = work_dir
         # Working directory - other subdirectories are created in this one
@@ -29,8 +66,6 @@ class SamplingPoolPBS(SamplingPool):
         # Weight of the single PBS script (putting more small jobs into single PBS job).
         self._current_job_weight = 0
         # Current collected weight.
-        self._job_count = job_count
-        # Current number of jobs - sort of jobID
         self._n_samples_in_job = 0
         # Number of samples in job
         self.pbs_script = None
@@ -42,11 +77,18 @@ class SamplingPoolPBS(SamplingPool):
         # List of scheduled samples
         self._pbs_ids = []
         # List of pbs job ids which should run
+        self._unfinished_sample_ids = set()
+        # List of sample id which are not collected - collection attempts are done in the get_finished()
+
+        self.clean = clean
 
         self._output_dir = None
         self._jobs_dir = None
         self._create_output_dir()
         self._create_job_dir()
+
+        self._job_count = self._get_job_count()
+        # Current number of jobs - sort of jobID
 
     def _create_output_dir(self):
         """
@@ -55,7 +97,7 @@ class SamplingPoolPBS(SamplingPool):
         """
         self._output_dir = os.path.join(self._work_dir, SamplingPoolPBS.OUTPUT_DIR)
 
-        if os.path.isdir(self._output_dir):
+        if self.clean and os.path.isdir(self._output_dir):
             shutil.rmtree(self._output_dir)
 
         os.makedirs(self._output_dir, mode=0o775, exist_ok=True)
@@ -67,6 +109,19 @@ class SamplingPoolPBS(SamplingPool):
         """
         self._jobs_dir = os.path.join(self._output_dir, SamplingPoolPBS.JOBS_DIR)
         os.makedirs(self._jobs_dir, mode=0o775, exist_ok=True)
+
+    def _get_job_count(self):
+        """
+        Get number of created jobs.
+        :return:
+        """
+        files_pattern = os.path.join(self._jobs_dir, "*_job.sh")
+        files = glob.glob(files_pattern)
+        if not files:
+            return 0
+
+        job_id = re.findall(r'(\d+)_job.sh', files[-1])[0]
+        return int(job_id) + 1
 
     def _save_structure(self):
         """
@@ -108,12 +163,8 @@ class SamplingPoolPBS(SamplingPool):
                                      '#PBS -e {pbs_output_dir}/{job_name}.ER',
                                      '']
 
-        self._pbs_header_template.extend(('cd {work_dir}',
-                                          'source env/bin/activate',
-                                          'pip3 install /storage/liberec3-tul/home/martin_spetlik/MLMC_new_design',))
-
-        self._pbs_header_template.extend(kwargs['modules'])
-        self._pbs_header_template.extend(('{python} {pbs_process_file_dir}/pbs_process.py {output_dir} {job_name} >'
+        self._pbs_header_template.extend(kwargs['env_setting'])
+        self._pbs_header_template.extend(('{python} -m mlmc.pbs_job {output_dir} {job_name} >'
                                           '{pbs_output_dir}/{job_name}_STDOUT 2>&1',))
         self._pbs_config = kwargs
 
@@ -154,8 +205,8 @@ class SamplingPoolPBS(SamplingPool):
         if len(self._scheduled) > 0:
             job_id = "{:04d}".format(self._job_count)
             # Create pbs job
-            pbs_process = PbsProcess.create_job(self._output_dir, self._jobs_dir, job_id,
-                                                                       SamplingPoolPBS.LEVEL_SIM_CONFIG)
+            pbs_process = PbsJob.create_job(self._output_dir, self._jobs_dir, job_id,
+                                            SamplingPoolPBS.LEVEL_SIM_CONFIG)
             # Write scheduled samples to file
             pbs_process.save_scheduled(self._scheduled)
 
@@ -168,7 +219,10 @@ class SamplingPoolPBS(SamplingPool):
             # Write pbs script
             job_file = os.path.join(self._jobs_dir, SamplingPoolPBS.JOB.format(job_id))
             script_content = "\n".join(self.pbs_script)
+
             self.write_script(script_content, job_file)
+            self.write_script(str(self._job_count), SamplingPoolPBS.JOBS_COUNT)
+            # Write current job count
             self._job_count += 1
 
             #subprocess.call(job_file)
@@ -218,7 +272,7 @@ class SamplingPoolPBS(SamplingPool):
         """
         self.execute()
         finished_pbs_jobs, unfinished_pbs_jobs = self._qstat_pbs_job()
-        return self.get_result_files(finished_pbs_jobs, unfinished_pbs_jobs)
+        return self._get_result_files(finished_pbs_jobs, unfinished_pbs_jobs)
 
     def _qstat_pbs_job(self):
         """
@@ -249,7 +303,7 @@ class SamplingPoolPBS(SamplingPool):
 
         return finished_pbs_jobs, unfinished_pbs_jobs
 
-    def get_result_files(self, finished_pbs_jobs, unfinished_pbs_jobs):
+    def _get_result_files(self, finished_pbs_jobs, unfinished_pbs_jobs):
         """
         Get results from files
         :param finished_pbs_jobs: List[str], finished pbs jobs,
@@ -267,7 +321,7 @@ class SamplingPoolPBS(SamplingPool):
             reg = "*_{}".format(pbs_id)
             file = glob.glob(reg)
             job_id = re.findall(r'(\d+)_\d+', file[0])[0]
-            n_running += PbsProcess.get_job_n_running(job_id, self._jobs_dir)
+            n_running += PbsJob.get_job_n_running(job_id, self._jobs_dir)
 
         successful_results = {}
         failed_results = {}
@@ -281,28 +335,87 @@ class SamplingPoolPBS(SamplingPool):
                 file = file[0]
                 job_id = re.findall(r'(\d+)_\d+', file)[0]
                 # Get sample results
-                successful, failed, time = PbsProcess.read_results(job_id, self._jobs_dir)
+                successful, failed, time = PbsJob.read_results(job_id, self._jobs_dir)
 
                 # Split results to levels
                 for level_id, results in successful.items():
                     successful_results.setdefault(level_id, []).extend(results)
-                for level_id, results in failed_results.items():
+                for level_id, results in failed.items():
                     failed_results.setdefault(level_id, []).extend(results)
-                for level_id, results in times.items():
-                    times[level_id] = results
+                for level_id, results in time.items():
+                    times[level_id] = results[-1]
 
                 # Delete pbsID file - it means job is finished
                 SamplingPoolPBS.delete_pbs_id_file(file)
 
-        # @TODO: get sample ids for failed pbs jobs and add these samples to failed ones with particular error message
+        if self._unfinished_sample_ids:
+            successful_results, failed_results, times = self._collect_unfinished(successful_results,
+                                                                                 failed_results, times)
 
         return successful_results, failed_results, n_running, times
-
-    def have_permanent_sample(self, sample_id):
+    
+    def _load_sample_id_job_id_map(self, sample_dir):
         """
-        List of permanent samples or find per call?
+        :param sample_dir: path to sample directory
+        :return: str, job id
         """
+        file_name = os.path.join(sample_dir, PbsJob.PERMANENT_SAMPLE.format("*"))
+        file = glob.glob(file_name)[0]
+        job_id = re.findall(r'._(\d+)', file)[0]
 
+        return job_id
+
+    def _collect_unfinished(self, successful_results, failed_results, times):
+        """
+        Collect samples which had finished after main process crashed, append them to new collected samples
+        :param successful_results: dict
+        :param failed_results: dict
+        :param times: dict
+        :return: all input dictionaries
+        """
+        already_collected = set()
+
+        for sample_id in self._unfinished_sample_ids:
+            if sample_id in already_collected:
+                continue
+
+            sample_dir = os.path.join(self._output_dir, sample_id)
+            job_id = self._load_sample_id_job_id_map(sample_dir)
+
+            successful, failed, time = PbsJob.read_results(job_id, self._jobs_dir)
+
+            # Split results to levels
+            for level_id, results in successful.items():
+                for res in results:
+                    if res[0] in self._unfinished_sample_ids:
+                        already_collected.add(res[0])
+                        successful_results.setdefault(level_id, []).append(res)
+
+            for level_id, results in failed_results.items():
+                for res in results:
+                    if res[0] in self._unfinished_sample_ids:
+                        already_collected.add(res[0])
+                        failed_results.setdefault(level_id, []).append(res)
+
+            for level_id, results in times.items():
+                for res in results:
+                    if res[0] in self._unfinished_sample_ids:
+                        times.setdefault(level_id, []).append(res)
+                times[level_id] = results
+
+            # Delete pbsID file - it means job is finished
+            # SamplingPoolPBS.delete_pbs_id_file(file)
+
+        self._unfinished_sample_ids = set()
+
+        return successful_results, failed_results, times
+
+    def have_permanent_samples(self, sample_ids):
+        """
+        List of unfinished sample ids, the corresponding samples are collecting in next get_finished() call .
+        """
+        self._unfinished_sample_ids = set(sample_ids)
+        
     @staticmethod
     def delete_pbs_id_file(file_path):
         """
